[quantizer] enable fbgemm & tests (#41)
* [quantizer] enable fbgemm

* [CI] enable tests for fbgemm on windows

* [CI] swap -> 8GB

* [converter] qconv without bias

* [CI] enable quantized-op tests

* [tests] ci-only logic

* [CI] restore

+++ tests/converter_qat_test.py
@@ -15,7 +15,7 @@ import numpy as np
 from tinynn.converter import TFLiteConverter
 from tinynn.graph.tracer import model_tracer, trace
 from tinynn.graph.quantization.quantizer import QATQuantizer
-from common_utils import collect_custom_models, collect_torchvision_models, prepare_inputs
+from common_utils import IS_CI, collect_custom_models, collect_torchvision_models, prepare_inputs
 
 
 HAS_TF = False
@@ -112,7 +112,11 @@ class TestModelMeta(type):
 
                 inputs = prepare_inputs(m)
 
-                quantizer = QATQuantizer(m, inputs, work_dir='out', config={'remove_weights_after_load': True})
+                config = {'remove_weights_after_load': True}
+                if sys.platform == 'win32':
+                    config.update({'backend': 'fbgemm', 'per_tensor': False})
+
+                quantizer = QATQuantizer(m, inputs, work_dir='out', config=config)
                 qat_model = quantizer.quantize()
 
             return qat_model, inputs
@@ -137,13 +141,19 @@ class TestModelMeta(type):
                 qat_model = torch.quantization.convert(qat_model)
 
                 out_path = f'out/{model_file}.tflite'
-                out_pt = f'out/{model_file}.pt'
-                converter = TFLiteConverter(
-                    qat_model, inputs, out_path, dump_jit_model_path=out_pt, gc_when_reload=True
-                )
+
+                extra_kwargs = {}
+                if IS_CI:
+                    out_pt = f'out/{model_file}.pt'
+                    extra_kwargs.update({'dump_jit_model_path': out_pt, 'gc_when_reload': True})
+                if sys.platform == 'win32':
+                    extra_kwargs.update({'quantize_target_type': 'int8'})
+
+                converter = TFLiteConverter(qat_model, inputs, out_path, **extra_kwargs)
                 converter.convert()
 
-                os.remove(out_pt)
+                if IS_CI:
+                    os.remove(out_pt)
 
                 if HAS_TF:
                     outputs = converter.get_outputs()
@@ -152,10 +162,12 @@ class TestModelMeta(type):
                     tf_outputs = get_tflite_out(out_path, input_tf)
                     self.assertTrue(len(outputs) == len(tf_outputs))
 
+                if IS_CI and os.path.exists(out_path):
+                    os.remove(out_path)
+
         return f
 
 
-@unittest.skipIf(sys.platform == 'win32', 'Quantization cannot be performed on Windows')
 class TestModel(unittest.TestCase, metaclass=TestModelMeta):
     pass
